chore: adjust the code
