+++ projects/app/.env.template
@@ -102,6 +102,7 @@ CONFIG_JSON_PATH=
 
 
 # S3 Config
+# S3_CUSTOM_ENDPOINT=http://s3.example.com/bucketname
 S3_ENDPOINT=localhost
 S3_PORT=9000
 S3_USE_SSL=false
