+++ projects/app/src/pageComponents/dataset/detail/CollectionCard/BackupImportModal.tsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import MyModal from '@fastgpt/web/components/common/MyModal';
 import { useTranslation } from 'next-i18next';
 import { Box, Button, HStack, ModalBody, ModalFooter, VStack } from '@chakra-ui/react';
-import FileSelector, { type SelectFileItemType } from '../components/FileSelector';
 import MyIcon from '@fastgpt/web/components/common/Icon';
 import MyIconButton from '@fastgpt/web/components/common/Icon/button';
 import { postBackupDatasetCollection } from '@/web/core/dataset/api';
@@ -10,6 +9,7 @@ import { useRequest2 } from '@fastgpt/web/hooks/useRequest';
 import { DatasetPageContext } from '@/web/core/dataset/context/datasetPageContext';
 import { useContextSelector } from 'use-context-selector';
 import LightTip from '@fastgpt/web/components/common/LightTip';
+import FileSelectorBox, { type SelectFileItemType } from '@/components/Select/FileSelectorBox';
 
 const BackupImportModal = ({
   onFinish,
@@ -46,7 +46,7 @@ const BackupImportModal = ({
       <ModalBody>
         <LightTip mb={3} icon="common/info" text={t('dataset:backup_dataset_tip')} />
 
-        <FileSelector
+        <FileSelectorBox
           maxCount={1}
           fileType=".csv"
           selectFiles={selectFiles}
