+++ projects/app/src/pages/api/plugin/delete.ts
@@ -1,25 +1,17 @@
-import type { NextApiRequest, NextApiResponse } from 'next';
+import type { NextApiResponse } from 'next';
 import { NextAPI } from '@/service/middleware/entry';
 import { deleteSystemTool } from '@fastgpt/service/core/app/tool/api';
+import type { ApiRequestProps } from '@fastgpt/service/type/next';
+import { PluginSourceEnum } from '@fastgpt/global/core/app/plugin/constants';
 
-async function handler(req: NextApiRequest, res: NextApiResponse<any>) {
-  const toolId = (req.query.toolId as string) || req.body?.toolId;
+async function handler(req: ApiRequestProps<{}, { toolId: string }>, res: NextApiResponse<any>) {
+  const { toolId } = req.query;
 
   if (!toolId) {
     return Promise.reject('ToolId is required');
   }
 
-  const actualToolId = toolId.includes('-') ? toolId.split('-').slice(1).join('-') : toolId;
-
-  const result = await deleteSystemTool(actualToolId);
-
-  return result;
+  return deleteSystemTool(toolId.replace(`${PluginSourceEnum.systemTool}-`, ''));
 }
 
 export default NextAPI(handler);
-
-export const config = {
-  api: {
-    bodyParser: true
-  }
-};
