+++ projects/app/src/web/common/file/api.ts
@@ -32,15 +32,3 @@ export const postS3UploadFile = (
     },
     onUploadProgress
   });
-export const getPluginUploadPresignedURL = (data: { filename: string }) =>
-  GET<UploadPresignedURLResponse>('/common/file/plugin/getUploadURL', data);
-
-export const postConfirmUpload = (data: { objectName: string }) =>
-  GET<string>('/common/file/plugin/confirmUpload', data);
-
-// export const postUploadFileAndUrl = (url: string) => POST('/plugin/upload', { url });
-
-export const postDeletePlugin = (toolId: string) =>
-  DELETE('/plugin/delete', {
-    toolId
-  });
